ZXP-108 Move AvisService and resolve company ID issue for avi
