+++ front-angular/src/app/components/entreprises/detail-company/detail-company.component.html
@@ -1,17 +1,17 @@
 <div>
   <div class="banner">
-    <h2><img class="company-img" src="{{company.srcImage}}" alt="logo company" /></h2>
-    <h2>{{company.name}}</h2><br />
+    <h2><img class="company-img" src="{{company?.srcImage}}" alt="logo company" /></h2>
+    <h2>{{company?.name}}</h2><br />
   </div>
   <div class="background-img">
     <img style="width:100%;" src="../../../../assets/images/background-test.jpg" alt="banner" />
   </div>
   <div class="offer-sheet">
-    <p><span class="label">Date de création : </span>{{company.creationDate}}</p>
-    <p><span class="label">Localisation : </span>{{company.location}}</p>
-    <p><span class="label">Taille de l'entreprise : </span>{{company.taille}}</p>
-    <p><span class="label">Description : </span>{{company.description}}</p>
-    <p><span class="label">Contact : </span>{{company.contact}}</p>
+    <p><span class="label">Date de création : </span>{{company?.creationDate}}</p>
+    <p><span class="label">Localisation : </span>{{company?.location}}</p>
+    <p><span class="label">Taille de l'entreprise : </span>{{company?.taille}}</p>
+    <p><span class="label">Description : </span>{{company?.description}}</p>
+    <p><span class="label">Contact : </span>{{company?.contact}}</p>
   </div>
 </div>
 
@@ -19,5 +19,5 @@
 
 <div class="avis">
   <h1>Avis</h1>
-  <app-avis-company [idCompany]="company._id"></app-avis-company>
+  <app-avis-company [idCompany]="companyId"></app-avis-company>
 </div>
