+++ front-angular/src/app/services/index.ts
@@ -1,3 +1,4 @@
 export * from './authentication.service';
 export * from './company.service';
 export * from './user.service';
+export * from './avis.service';
